Convert to a command
